Added configure script for ffmpeg extension, and the resultant libraries.
Note: Normally we wouldn't commit binaries like this, but we've had
issues with these not being built before being pulled in as a depedency.
In the future we should integrate the ffmpeg build into the gradle build
process, but for now at least the binaries will be available without
needing to mess with JNI/NDK stuff every time.